fixed typos

+++ include/sht4x.h
@@ -97,121 +97,121 @@ esp_err_t sht4x_init(sht4x_t *const me, i2c_bus_t *i2c_bus, uint8_t dev_addr,
  * @brief Function for a single shot measurement with high repeatability.
  *
  * @param me          : Pointer to a sht4x_t instance
- * @param temperature : Temperature in milli degrees centigrade.
- * @param humidity    : Humidity in milli percent relative humidity.
+ * @param temperature : Temperature in degrees centigrade.
+ * @param humidity    : Humidity in percent relative humidity.
  *
  * @return 0 on success, an error code otherwise
  */
-esp_err_t sht4x_measure_high_precision(sht4x_t *const me, int32_t *temperature,
-		                                   int32_t *humidity);
+esp_err_t sht4x_measure_high_precision(sht4x_t *const me, float *temperature,
+		                                   float *humidity);
 
 /**
  * @brief Function for a single shot measurement with medium repeatability.
  *
  * @param me          : Pointer to a sht4x_t instance
- * @param temperature : Temperature in milli degrees centigrade.
- * @param humidity    : Humidity in milli percent relative humidity.
+ * @param temperature : Temperature in degrees centigrade.
+ * @param humidity    : Humidity in percent relative humidity.
  *
  * @return 0 on success, an error code otherwise
  */
-esp_err_t sht4x_measure_medium_precision(sht4x_t *const me, int32_t *temperature,
-		                                     int32_t *humidity);
+esp_err_t sht4x_measure_medium_precision(sht4x_t *const me, float *temperature,
+		                                     float *humidity);
 
 /**
  * @brief Function for a single shot measurement with lowest repeatability.
  *
  * @param me          : Pointer to a sht4x_t instance
- * @param temperature : Temperature in milli degrees centigrade.
- * @param humidity    : Humidity in milli percent relative humidity.
+ * @param temperature : Temperature in degrees centigrade.
+ * @param humidity    : Humidity in percent relative humidity.
  *
  * @return 0 on success, an error code otherwise
  */
-esp_err_t sht4x_measure_lowest_precision(sht4x_t *const me, int32_t *temperature,
-		                                     int32_t *humidity);
+esp_err_t sht4x_measure_lowest_precision(sht4x_t *const me, float *temperature,
+		                                     float *humidity);
 
 /**
  * @brief Function to activate highest heater power and perform a single
  * shot high precision measurement for 1s.
  *
  * @param me          : Pointer to a sht4x_t instance
- * @param temperature : Temperature in milli degrees centigrade.
- * @param humidity    : Humidity in milli percent relative humidity.
+ * @param temperature : Temperature in degrees centigrade.
+ * @param humidity    : Humidity in percent relative humidity.
  *
  * @return 0 on success, an error code otherwise
  */
 esp_err_t sht4x_activate_highest_heater_power_long(sht4x_t *const me,
-		                                               int32_t *temperature,
-																									 int32_t *humidity);
+		                                               float *temperature,
+																									 float *humidity);
 
 /**
  * @brief Function to activate highest heater power and perform a single
  * shot high precision measurement for 0.1s.
  *
  * @param me          : Pointer to a sht4x_t instance
- * @param temperature : Temperature in milli degrees centigrade.
- * @param humidity    : Humidity in milli percent relative humidity.
+ * @param temperature : Temperature in degrees centigrade.
+ * @param humidity    : Humidity in percent relative humidity.
  *
  * @return 0 on success, an error code otherwise
  */
 esp_err_t sht4x_activate_highest_heater_power_short(sht4x_t *const me,
-		                                                int32_t *temperature,
-																										int32_t *humidity);
+		                                                float *temperature,
+																										float *humidity);
 
 /**
  * @brief Function to activate medium heater power and perform a single
  * shot high precision measurement for 1s.
  *
  * @param me          : Pointer to a sht4x_t instance
- * @param temperature : Temperature in milli degrees centigrade.
- * @param humidity    : Humidity in milli percent relative humidity.
+ * @param temperature : Temperature in degrees centigrade.
+ * @param humidity    : Humidity in percent relative humidity.
  *
  * @return 0 on success, an error code otherwise
  */
 esp_err_t sht4x_activate_medium_heater_power_long(sht4x_t *const me,
-		                                              int32_t *temperature,
-																									int32_t *humidity);
+		                                              float *temperature,
+																									float *humidity);
 
 /**
  * @brief Function to activate medium heater power and perform a single
  * shot high precision measurement for 0.1s.
  *
  * @param me          : Pointer to a sht4x_t instance
- * @param temperature : Temperature in milli degrees centigrade.
- * @param humidity    : Humidity in milli percent relative humidity.
+ * @param temperature : Temperature in degrees centigrade.
+ * @param humidity    : Humidity in percent relative humidity.
  *
  * @return 0 on success, an error code otherwise
  */
 esp_err_t sht4x_activate_medium_heater_power_short(sht4x_t *const me,
-		                                               int32_t *temperature,
-																									 int32_t *humidity);
+		                                               float *temperature,
+																									 float *humidity);
 
 /**
  * @brief Function to activate lowest heater power and perform a single
  * shot high precision measurement for 1s.
  *
  * @param me          : Pointer to a sht4x_t instance
- * @param temperature : Temperature in milli degrees centigrade.
- * @param humidity    : Humidity in milli percent relative humidity.
+ * @param temperature : Temperature in degrees centigrade.
+ * @param humidity    : Humidity in percent relative humidity.
  *
  * @return 0 on success, an error code otherwise
  */
 esp_err_t sht4x_activate_lowest_heater_power_long(sht4x_t *const me,
-		                                              int32_t *temperature,
-																									int32_t *humidity);
+		                                              float *temperature,
+																									float *humidity);
 
 /**
  * @brief Function to activate lowest heater power and perform a single
  * shot high precision measurement for 0.1s.
  *
  * @param me          : Pointer to a sht4x_t instance
- * @param temperature : Temperature in milli degrees centigrade.
- * @param humidity    : Humidity in milli percent relative humidity.
+ * @param temperature : Temperature in degrees centigrade.
+ * @param humidity    : Humidity in percent relative humidity.
  *
  * @return 0 on success, an error code otherwise
  */
 esp_err_t sht4x_activate_lowest_heater_power_short(sht4x_t *const me,
-		                                               int32_t *temperature,
-																									 int32_t *humidity);
+		                                               float *temperature,
+																									 float *humidity);
 
 /**
  * @brief Function for a single shot measurement with high repeatability.
